refactor: cart remove button
