+++ src/advanced/components/cart/item.tsx
@@ -1,15 +1,13 @@
 import { Product } from 'src/advanced/data/products';
 import { CartItemSummary } from './summary';
-import { useProductList } from '../../context/product';
 import { CartQuantityButton } from './quantity-button';
+import { CartRemoveButton } from './remove-button';
 
 interface CartItemProps extends React.HTMLAttributes<HTMLDivElement> {
   cart: Product;
 }
 
 export const CartItem: React.FC<CartItemProps> = ({ cart, ...props }) => {
-  const { cartList, setProductList } = useProductList();
-
   return (
     <div
       key={cart.id}
@@ -21,31 +19,7 @@ export const CartItem: React.FC<CartItemProps> = ({ cart, ...props }) => {
       <div>
         <CartQuantityButton step={-1} label="-" cart={cart} />
         <CartQuantityButton step={1} label="+" cart={cart} />
-        <button
-          className="remove-item bg-red-500 text-white px-2 py-1 rounded"
-          data-product-id={cart.id}
-          onClick={() => {
-            const productToRemove = cartList.find(
-              (product) => product.id === cart.id,
-            );
-            if (productToRemove) {
-              setProductList((prev) =>
-                prev.map((product) =>
-                  product.id === productToRemove.id
-                    ? {
-                        ...product,
-                        cartQuantity: 0,
-                        stockQuantity:
-                          product.stockQuantity + product.cartQuantity,
-                      }
-                    : product,
-                ),
-              );
-            }
-          }}
-        >
-          삭제
-        </button>
+        <CartRemoveButton cart={cart} />
       </div>
     </div>
   );
